third_party/WebKit/LayoutTests/bluetooth/getPrimaryService/delayed-discovery-no-permission-absent-service.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let expected = new DOMException('Origin is not allowed to access the ' +
                                  'service. Remember to add the service to a ' +
                                  'filter or to optionalServices in ' +
                                  'requestDevice().',
                                  'SecurityError');
  return setBluetoothFakeAdapter('DelayedServicesDiscoveryAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => Promise.all([
      assert_promise_rejects_with_message(
        gattServer.getPrimaryService(glucose.alias), expected),
      assert_promise_rejects_with_message(
        gattServer.getPrimaryService(glucose.name), expected),
      assert_promise_rejects_with_message(
        gattServer.getPrimaryService(glucose.uuid), expected)]));
}, 'Delayed services discovered, request for absent service without ' +
   'permission. Reject with SecurityError.');
</script>
